+++ lib/src/states/GetHelperState.dart
@@ -16,21 +16,28 @@ class GetHelperState extends StubbleState {
     _helper = '';
     _attributes = [];
 
-    return StubbleResult(state: GetBlockNameState());
+    return StubbleResult(
+      state: GetBlockNameState(),
+    );
   }
 
   StubbleResult process(ProcessMessage msg, StubbleContext context) {
     final charCode = msg.charCode;
 
     if (charCode == CLOSE_BRACKET) {
-      return StubbleResult(state: CloseBracketState());
+      return StubbleResult(
+        state: CloseBracketState(),
+      );
     } else if (charCode == SPACE) {
       return null;
     }
 
     return StubbleResult(
-        state: GetAttributeState(),
-        message: ProcessMessage(charCode: charCode));
+      state: GetAttributeState(),
+      message: ProcessMessage(
+        charCode: charCode,
+      ),
+    );
   }
 
   StubbleResult notify(NotifyMessage msg, StubbleContext context) {
@@ -38,16 +45,22 @@ class GetHelperState extends StubbleState {
       case NOTIFY_NAME_RESULT:
         _helper = msg.value;
 
-        if (!context.callable(_helper)) {}
-
-        return StubbleResult(message: ProcessMessage(charCode: msg.charCode));
+        return StubbleResult(
+          message: ProcessMessage(
+            charCode: msg.charCode,
+          ),
+        );
       case NOTIFY_SECOND_CLOSE_BRACKET_FOUND:
         return result(context);
       case NOTIFY_ATTR_RESULT:
         _attributes.add(msg.value);
 
         if (msg.charCode != null) {
-          return StubbleResult(message: ProcessMessage(charCode: msg.charCode));
+          return StubbleResult(
+            message: ProcessMessage(
+              charCode: msg.charCode,
+            ),
+          );
         }
 
         break;
@@ -59,6 +72,14 @@ class GetHelperState extends StubbleState {
   }
 
   StubbleResult result(StubbleContext context) {
+    if (!context.callable(_helper)) {
+      return StubbleResult(
+          pop: true,
+          err: StubbleError(
+              code: ERROR_HELPER_UNREGISTERED,
+              text: 'Helper "$_helper" is unregistered'));
+    }
+
     final result = StubbleResult();
 
     try {
@@ -69,6 +90,7 @@ class GetHelperState extends StubbleState {
       result.result = context.call(_helper, _attributes, null);
       result.pop = true;
     } catch (e) {
+      result.pop = true;
       result.err = StubbleError(
           text: 'Error in helper function ${_helper}: ${e.toString()}',
           code: ERROR_CALLING_HELPER);