+++ lib/src/states/GetWithBlockState.dart
@@ -1,10 +1,13 @@
 part of stubble;
 
 class GetWithBlockState extends StubbleState {
-  String _path;
-  String _body;
+  final int symbol;
+  final int line;
 
-  GetWithBlockState() {
+  String _path = '';
+  String _body = '';
+
+  GetWithBlockState({this.symbol, this.line}) {
     methods = {
       'process': (msg, context) => process(msg, context),
       'notify': (msg, context) => notify(msg, context),
@@ -14,35 +17,49 @@ class GetWithBlockState extends StubbleState {
   StubbleResult process(ProcessMessage msg, StubbleContext context) {
     final charCode = msg.charCode;
 
-    if (charCode == CLOSE_BRACKET) {
-      return StubbleResult(state: CloseBracketState());
+    if (charCode == EOS) {
+      return StubbleResult(
+        err: StubbleError(
+          code: ERROR_UNTERMINATED_BLOCK,
+          text: 'Unterminated "WITH" block at $line:$symbol',
+        ),
+      );
+    } else if (charCode == CLOSE_BRACKET) {
+      return StubbleResult(
+        state: CloseBracketState(),
+      );
     } else if (charCode == SPACE) {
       return null;
     }
 
     return StubbleResult(
-        state: GetPathState(), message: ProcessMessage(charCode: charCode));
+      state: GetPathState(),
+      message: ProcessMessage(
+        charCode: charCode,
+      ),
+    );
   }
 
   StubbleResult notify(NotifyMessage msg, StubbleContext context) {
     switch (msg.type) {
       case NOTIFY_PATH_RESULT:
         _path = msg.value;
-        return StubbleResult(message: ProcessMessage(charCode: msg.charCode));
+        return StubbleResult(
+          message: ProcessMessage(
+            charCode: msg.charCode,
+          ),
+        );
 
       case NOTIFY_SECOND_CLOSE_BRACKET_FOUND:
-        return StubbleResult(state: GetBlockEndState(blockName: 'with'));
+        return StubbleResult(
+          state: GetBlockEndState(
+            blockName: 'with',
+          ),
+        );
 
       case NOTIFY_BLOCK_END_RESULT:
         _body = msg.value;
         return result(context);
-
-      default:
-        return StubbleResult(
-            err: StubbleError(
-                code: ERROR_UNSUPPORTED_NOTIFY,
-                text:
-                    'State "$runtimeType" does not support notifies of type ${msg.type}'));
     }
   }
 
@@ -51,7 +68,9 @@ class GetWithBlockState extends StubbleState {
 
     if (_path == null || _path.isEmpty) {
       result.err = StubbleError(
-          text: 'With block required path to context data', code: ERROR_PATH_NOT_SPECIFIED);
+        text: 'With block required path to context data',
+        code: ERROR_PATH_NOT_SPECIFIED,
+      );
     } else {
       try {
         final data = context.get(_path);
@@ -66,17 +85,21 @@ class GetWithBlockState extends StubbleState {
             );
           } else {
             result.err = StubbleError(
-                text: '"With" block data should have "Map" type',
-                code: ERROR_WITH_DATA_MALFORMED);
+              text: '"With" block data should have "Map" type',
+              code: ERROR_WITH_DATA_MALFORMED,
+            );
           }
         } else {
           result.err = StubbleError(
-              text: 'Can\'t get data from context by path "${_path}"',
-              code: ERROR_PATH_WRONG_SPECIFIED);
+            text: 'Can\'t get data from context by path "${_path}"',
+            code: ERROR_PATH_WRONG_SPECIFIED,
+          );
         }
       } catch (e) {
-        result.err =
-            StubbleError(text: e.toString(), code: ERROR_CALLING_HELPER);
+        result.err = StubbleError(
+          text: e.toString(),
+          code: ERROR_CALLING_HELPER,
+        );
       }
     }