+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 1.0.0
 
-- Initial version, created by Stagehand
+- Initial version
+
+## 1.0.1
+
+- New error types added.
+- Stubble options can be passed to `Stubble()` by the second parameter:
+
+Available options:
+    - `ignoreUnregisteredHelperErrors` - ignore errors of 'Unregistered helpers'
+    - `ignoreTagCaseSensetive` - ignore case of open and close tags